Added an Initializer to create databases and collections if they don't exist
Normally MongoDB will automatically create databases and collections if they are missing. But, it will only create the default collection where there are two collections in one. So when it makes a query, it will have to travel through two collections to actually find the value. Which will slow down the query speed. To avoid this problem, MongoDB has a collection called a clustered collection. But MongoDB will not create this by default, So we have to create it manually. The Initializer checks if all the databases and collections exist, if not it will create the missing database and collection as a clustered collection.